pyproject.toml
[project]
name = "agent-zero"
version = "0.1.0"
requires-python = ">=3.12"
description = "Agent-based decarbonisation pathways model"
license = "MIT"
readme = "README.md"
authors = [
  {name = "dlg0"}
]
keywords = ["agent-based", "modeling", "decarbonisation", "simulation"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
]
dependencies = [
  "pandas>=2.2",
  "pyarrow>=16.0",
  "numpy>=2.0",
  "pyyaml>=6.0",
  "click>=8.1",
  "rich>=13.0",
  "rich-click>=1.8"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=4.0",
  "ruff>=0.8",
  "mypy>=1.13",
  "pandas-stubs>=2.2",
  "types-PyYAML>=6.0",
]
story = [
  "anthropic>=0.39.0",
]

[project.urls]
Homepage = "https://github.com/dlg0/agent-zero"
Repository = "https://github.com/dlg0/agent-zero"
Issues = "https://github.com/dlg0/agent-zero/issues"

[project.scripts]
agentzero = "agent_zero.cli:main"
az = "agent_zero.cli:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra -q"
xfail_strict = true
markers = [
    "integration: tests that touch disk / real data / multiple modules",
    "e2e: full end-to-end runs (slowest)",
]

[tool.coverage.run]
source = ["agent_zero"]
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

# ---------------------------------------------------------------------------
# Ruff (linting + formatting)
# ---------------------------------------------------------------------------
[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "RUF",    # Ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # function call in default argument (Click uses this)
    "TCH",    # type-checking imports (overly strict for this project)
    "RUF001", # ambiguous unicode characters in strings (intentional for CLI icons)
    "RUF002", # ambiguous unicode characters in docstrings
    "RUF003", # ambiguous unicode characters in comments
]

[tool.ruff.lint.isort]
known-first-party = ["agent_zero"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

# ---------------------------------------------------------------------------
# Mypy (type checking)
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.12"
strict = false
warn_return_any = false
warn_unused_ignores = true
show_error_codes = true
check_untyped_defs = true
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = ["pyarrow", "numpy", "click"]
ignore_missing_imports = true